add hashed fetch, update logic
