+++ assets/scripts/react/components/index.jsx
@@ -1,5 +1,4 @@
-import GithubSignInButton from "./GithubSignInButton";
 import AuthPage from "./AuthPage";
 import CloneButton from "./CloneButton";
 
-export default [GithubSignInButton, AuthPage, CloneButton];
+export default [AuthPage, CloneButton];
